Seperate construction logic
This commit is a preparation activity of breaking up the super large
- Creating a new site moves to the ConstructionManager
- Placement of the site moves to BuildingPlacement

Changes:
- Create a new BuildingPlacement class that is responsible for placing a new Site
- Logic is moved from ConstructionMission into BuildingPlacement
- Create basic unit test
- Creation and determination of the new site is moved to the ConstructionManager
- Update ConstructionManager unit test

In addition
- Unify the tooltip creation for ConstructionStage into a ConstructionStageFormat
- Applied to ConstructionPanel & ConstructionSitesPanel

Ref #1642
