+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/mission/ConstructionMission.java
@@ -6,8 +6,6 @@
  */
 package com.mars_sim.core.person.ai.mission;
 
-import java.awt.geom.Line2D;
-import java.awt.geom.Point2D;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -17,23 +15,16 @@ import java.util.List;
 import java.util.Map;
 
 import com.mars_sim.core.LocalAreaUtil;
-import com.mars_sim.core.SimulationConfig;
 import com.mars_sim.core.UnitEventType;
 import com.mars_sim.core.UnitType;
 import com.mars_sim.core.building.Building;
-import com.mars_sim.core.building.BuildingConfig;
-import com.mars_sim.core.building.BuildingManager;
-import com.mars_sim.core.building.BuildingSpec;
 import com.mars_sim.core.building.construction.ConstructionSite;
 import com.mars_sim.core.building.construction.ConstructionStage;
 import com.mars_sim.core.building.construction.ConstructionStageInfo;
-import com.mars_sim.core.building.construction.ConstructionValues;
 import com.mars_sim.core.building.construction.ConstructionVehicleType;
-import com.mars_sim.core.building.function.FunctionType;
 import com.mars_sim.core.equipment.EVASuit;
 import com.mars_sim.core.equipment.EquipmentType;
 import com.mars_sim.core.logging.SimLogger;
-import com.mars_sim.core.map.location.BoundedObject;
 import com.mars_sim.core.map.location.LocalPosition;
 import com.mars_sim.core.mission.Construction;
 import com.mars_sim.core.mission.objectives.ConstructionObjective;
@@ -43,7 +34,6 @@ import com.mars_sim.core.person.ai.task.ConstructBuilding;
 import com.mars_sim.core.person.ai.task.util.Worker;
 import com.mars_sim.core.resource.ItemResourceUtil;
 import com.mars_sim.core.resource.Part;
-import com.mars_sim.core.structure.ObjectiveUtil;
 import com.mars_sim.core.structure.Settlement;
 import com.mars_sim.core.tool.RandomUtil;
 import com.mars_sim.core.vehicle.Crewable;
@@ -75,7 +65,6 @@ public class ConstructionMission extends AbstractMission
 	private static final MissionPhase CONSTRUCTION_PHASE = new MissionPhase("Mission.phase.construction");
 
 	private static final MissionStatus LUV_NOT_AVAILABLE = new MissionStatus("Mission.status.noLUV");
-	private static final MissionStatus CONSTRUCTION_SITE_NOT_FOUND_OR_CREATED =new MissionStatus("Mission.status.noConstructionSite");
 	private static final MissionStatus CONSTRUCTION_ENDED = new MissionStatus("Mission.status.constructionEnded");
 	private static final MissionStatus NEW_CONSTRUCTION_STAGE_NOT_DETERMINED = new MissionStatus("Mission.status.noConstructionStage");
 
@@ -83,34 +72,13 @@ public class ConstructionMission extends AbstractMission
 	public static final int MIN_PEOPLE = 2;
 	
 	private static final int MAX_PEOPLE = 10;
-
-
-	private static final int DIG_REGOLITH_PERCENT_PROBABILITY = 10;
 	
 	private static final int CONSTRUCT_PERCENT_PROBABILITY = 50;
 	
-	private static final double SMALL_AMOUNT = 0.001D;
 	/** Time (millisols) required to prepare construction site for stage. */
 	private static final double SITE_PREPARE_TIME = 250D;
-	// Default distance between buildings for construction.
-	private static final double DEFAULT_HABITABLE_BUILDING_DISTANCE = 5D;
-
-	private static final double DEFAULT_INHABITABLE_BUILDING_DISTANCE = 2D;
-
-	private static final double DEFAULT_FARMING_DISTANCE = 5D;
-
-	// Default width and length for variable size buildings if not otherwise
-	// determined.
-	private static final double DEFAULT_VARIABLE_BUILDING_WIDTH = 7D;
-
-	private static final double DEFAULT_VARIABLE_BUILDING_LENGTH = 9D;
-
-	/** Minimum length of a building connector (meters). */
-	private static final double MINIMUM_CONNECTOR_LENGTH = 1D;
 
 	private ConstructionObjective objective;
-
-	private static BuildingConfig buildingConfig = SimulationConfig.instance().getBuildingConfiguration();
 	
 	/**
 	 * Constructor 1 for Case 1: Determined by the need of the settlement.
@@ -129,7 +97,10 @@ public class ConstructionMission extends AbstractMission
 		setMissionCapacity(MAX_PEOPLE);
 
 		// Recruit additional members to mission.
-		recruitMembersForMission(startingMember, true, MIN_PEOPLE);
+		if (!recruitMembersForMission(startingMember, true, MIN_PEOPLE)) {
+			return;
+		}
+
 
 		// Determine construction site and stage.
 		int constructionSkill = 1;
@@ -139,32 +110,23 @@ public class ConstructionMission extends AbstractMission
 			person.getMind().setMission(this);
 		}
 
-		var site = determineSiteByProfit(startingMember.getAssociatedSettlement(), constructionSkill);
-		
-
-		if (!isDone()) {
-			// Set initial mission phase.
-			if (site == null) {
-				endMission(CONSTRUCTION_SITE_NOT_FOUND_OR_CREATED);
-				return;
-			}
-
-			else {
-				createObjectives(site, site.getCurrentConstructionStage(), null);
-			}
+		var site = startingMember.getAssociatedSettlement().getConstructionManager().getNextSite(constructionSkill);
+		if (site == null) {
+			endMission(NEW_CONSTRUCTION_STAGE_NOT_DETERMINED);
+			return;
 		}
+		createObjectives(site, site.getCurrentConstructionStage(), null);
 		
 		// Need to set the description of this mission correctly
 		// e.g. Pouring the foundation, Building the frame, or Constructing the building
 
-		// Create mission designation
-		createDesignationString();
 		// Call missionManager to add this mission
 	    missionManager.addMission(this);
 	}
 
 	private void createObjectives(ConstructionSite site, ConstructionStage stage, List<GroundVehicle> constructionVehicles) {
 		var settlement = site.getAssociatedSettlement();
+		site.setUndergoingConstruction(true);
 
 		// Reserve construction vehicles.
 		if (constructionVehicles == null) {
@@ -174,146 +136,12 @@ public class ConstructionMission extends AbstractMission
 		// Retrieve construction LUV attachment parts.
 		var luvAttachmentParts = retrieveConstructionLUVParts(settlement, stage, constructionVehicles);
 		
+		// Create mission designation
+		createDesignationString();
+
 		objective = new ConstructionObjective(site, stage, constructionVehicles, luvAttachmentParts);
 		addObjective(objective);
 		setPhase(PREPARE_SITE_PHASE, site.getAssociatedSettlement().getName());
-	}
-	/**
-	 * Determines the construction site based upon profit.
-	 * 
-	 * @param skill
-	 */
-	public ConstructionSite determineSiteByProfit(Settlement s, int skill) {
-
-		ConstructionSite site = null;
-		var manager = s.getConstructionManager();
-				
-		ConstructionValues values = manager.getConstructionValues();
-		values.clearCache();
-		double existingSitesProfit = values.getAllConstructionSitesProfit(skill);
-		double newSiteProfit = values.getNewConstructionSiteProfit(skill);
-		ConstructionStageInfo info = null;
-
-		logger.info(s, "existingSitesProfit: " + existingSitesProfit + "   newSiteProfit: " + newSiteProfit);
-		
-		if (existingSitesProfit > newSiteProfit) {
-			// If there are existing construction sites
-			logger.info(s, "Developing an existing construction site.");
-			
-			// Determine which existing construction site to work on.
-			double topSiteProfit = 0D;
-			Iterator<ConstructionSite> i = manager.getConstructionSitesNeedingConstructionMission().iterator();
-			while (i.hasNext()) {
-				ConstructionSite _site = i.next();
-				double siteProfit = values.getConstructionSiteProfit(_site, skill);
-				if (siteProfit > topSiteProfit) {
-					site = _site;
-					info = _site.getStageInfo();
-					topSiteProfit = siteProfit;
-				}
-			}
-			
-			site.setStageInfo(info);
-			
-			determineNewStage(site, info, skill, values);
-		}
-
-		else if (newSiteProfit >= 0D) {
-			// If there aren't any existing construction sites
-			logger.info(s, "Creating a new construction site.");
-			
-			// Case 1a: if using GUI			
-			// Case 1b: if not using GUI
-
-			// Create new site.
-			site = manager.createNewConstructionSite();
-
-			if (site == null)
-				logger.info(s, "site is null.");
-			
-			// Determine construction site new stage info via profits probability.
-			info = determineNewStageInfoByProfits(site, skill);
-
-			site.setStageInfo(info);
-			
-			// Determine construction site location and facing.
-			if (info != null) {
-				// Set construction site size.
-				if (info.getWidth() > 0D)
-					site.setWidth(info.getWidth());
-				else
-					// Set initial width value that may be modified later.
-					site.setWidth(DEFAULT_VARIABLE_BUILDING_WIDTH);
-
-				if (info.getLength() > 0D)
-					site.setLength(info.getLength());
-				else
-					// Set initial length value that may be modified later.
-					site.setLength(DEFAULT_VARIABLE_BUILDING_LENGTH);
-
-				determineNewStage(site, info, skill, values);
-
-				positionNewSite(site);
-
-				logger.info(s, "New construction site '" + site + "' added.");
-			}
-			
-			else {
-				logger.warning(s, "New construction stage could not be determined.");
-				endMission(NEW_CONSTRUCTION_STAGE_NOT_DETERMINED);
-				return null;
-			}
-		}
-		else {
-			logger.info(s, "Case 3");
-		}
-
-		return site;
-	}
-
-	/**
-	 * Determines a new stage to work on.
-	 * 
-	 * @param cSite
-	 * @param stageInfo
-	 * @param constructionSkill
-	 * @param values
-	 */
-	private void determineNewStage(ConstructionSite cSite, ConstructionStageInfo stageInfo, int constructionSkill,
-			ConstructionValues values) {
-		ConstructionStage stage = null;
-		var settlement = cSite.getAssociatedSettlement();
-		logger.info(settlement, "Determining a new stage to work on for " + cSite + ".");
-
-		// Determine new stage to work on.
-		if (cSite.hasUnfinishedStage()) {
-			stage = cSite.getCurrentConstructionStage();
-			logger.info(settlement, "Continuing work on existing site at " + settlement.getName());
-		}
-		
-		else {
-			logger.info(settlement, "Found no unfinished stages at " + cSite + ".");
-			
-			if (stageInfo == null) {
-				stageInfo = determineNewStageInfoByProfits(cSite, constructionSkill);
-			}
-
-			if (stageInfo != null) {
-				stage = new ConstructionStage(stageInfo, cSite);
-				cSite.addNewStage(stage);
-				values.clearCache();
-				logger.info(settlement, "Starting a new construction stage '" + stage + "' for " + cSite + ".");
-			} 
-			
-			else {
-				endMission(NEW_CONSTRUCTION_STAGE_NOT_DETERMINED);
-			}
-		}
-
-		// Mark site as undergoing construction.
-		if (stage != null) {
-			cSite.setUndergoingConstruction(true);
-		}
 	}
 
 	/**
@@ -341,68 +169,18 @@ public class ConstructionMission extends AbstractMission
 		
 		setMissionCapacity(MAX_PEOPLE);
 
-		ConstructionStage stage = null;
-		if (choosenSite != null) {
-			// site already selected
-			
-			// Case 2a: if using GUI	
-			// Case 2b: if not using GUI
-				
-			choosenSite.setStageInfo(stageInfo);
-
-			if (stageInfo != null) {
-				logger.info(settlement, "Case 2. stageInfo is " + stageInfo.getName());
-			} else {
-				logger.info(settlement, "Case 2. new construction stageInfo could not be determined.");
-			}
-	
-			stage = setupConstructionStage(choosenSite, stageInfo);
-
-			if (isDone()) {
-				return;
-			}
+		if (choosenSite == null) {
+			throw new IllegalArgumentException("Choosen site is missing");
 		}
 
-		else {
-			// site has NOT been selected
-			logger.info(settlement, "Case 3 : site has NOT been picked yet and the construction is manually started by users");
+		// site already selected
+		choosenSite.setStageInfo(stageInfo);
+		logger.info(settlement, "Case 2. new construction stageInfo could not be determined.");
 
-			// Case 3a: if using GUI to pick a site
-			// Case 3b: if GUI is NOT in use
+		var stage = setupConstructionStage(choosenSite, stageInfo);
 
-			var manager = settlement.getConstructionManager();
-
-			choosenSite = manager.createNewConstructionSite();
-
-			if (choosenSite != null) {
-
-				choosenSite.setStageInfo(stageInfo);
-				// Set construction site size.
-
-				if (stageInfo.getWidth() > 0D)
-					choosenSite.setWidth(stageInfo.getWidth());
-				else
-					// Set initial width value that may be modified later.
-					choosenSite.setWidth(DEFAULT_VARIABLE_BUILDING_WIDTH);
-
-				if (stageInfo.getLength() > 0D)
-					choosenSite.setLength(stageInfo.getLength());
-				else
-					// Set initial length value that may be modified later.
-					choosenSite.setLength(DEFAULT_VARIABLE_BUILDING_LENGTH);
-
-				positionNewSite(choosenSite);
-			}
-
-			else {
-				endMission(CONSTRUCTION_SITE_NOT_FOUND_OR_CREATED);
-				return;
-			}
-		
-			stage = setupConstructionStage(choosenSite, stageInfo);
-			if (isDone()) {
-				return;
-			}
+		if (isDone()) {
+			return;
 		}
 
 		createObjectives(choosenSite, stage, vehicles);
@@ -507,26 +285,7 @@ public class ConstructionMission extends AbstractMission
 		return luvAttachmentParts;
 	}
 
-	/**
-	 * Determines a new construction stage info for a site.
-	 *
-	 * @param site  the construction site.
-	 * @param skill the architect's construction skill.
-	 * @return construction stage info.
-	 * @throws Exception if error determining construction stage info.
-	 */
-	private ConstructionStageInfo determineNewStageInfoByProfits(ConstructionSite site, int skill) {
-		ConstructionStageInfo result = null;
-		var manager = site.getAssociatedSettlement().getConstructionManager();
-		ConstructionValues values = manager.getConstructionValues();
-		Map<ConstructionStageInfo, Double> stageProfits = values.getNewConstructionStageProfits(site, skill);
-		if (!stageProfits.isEmpty()) {
-			result = RandomUtil.getWeightedRandomObject(stageProfits);
-		}
-
-		return result;
-	}
-
+	
 	/**
 	 * Checks if a light utility vehicle (LUV) is available for the mission.
 	 *
@@ -583,7 +342,7 @@ public class ConstructionMission extends AbstractMission
 		if (SELECT_SITE_PHASE.equals(getPhase())) {
 			selectSitePhase();
 		} else if (PREPARE_SITE_PHASE.equals(getPhase())) {
-			prepareSitePhase(member, objective.getSite());
+			prepareSitePhase(objective.getSite());
 		} else if (CONSTRUCTION_PHASE.equals(getPhase())) {
 			constructionPhase(member);
 		}
@@ -616,12 +375,13 @@ public class ConstructionMission extends AbstractMission
 		}
 		return true;
 	}
+
 	/**
 	 * Performs the task in 'Prepares site' phase.
 	 *
 	 * @param member the mission member performing the phase.
 	 */
-	private void prepareSitePhase(Worker member, ConstructionSite site) {
+	private void prepareSitePhase(ConstructionSite site) {
 		if (!isPreReqsAvailable(site)) {
 			return;
 		}
@@ -780,17 +540,15 @@ public class ConstructionMission extends AbstractMission
 	public void endMission(MissionStatus endStatus) {
 		var site = objective.getSite();
 		// Mark site as not undergoing construction.
-		if (site != null)
-			site.setUndergoingConstruction(false);
+		site.setUndergoingConstruction(false);
 
 		// Unreserve all LUV attachment parts for this mission.
 		unreserveLUVparts(objective.getLuvAttachmentParts(), site.getAssociatedSettlement());
 
-		for (GroundVehicle v : objective.getConstructionVehicles()) {
-			if (v.getMission().equals(this)) {
-				v.setMission(null);
-			}
-		}
+		objective.getConstructionVehicles().stream()
+			.filter(v -> v.getMission().equals(this))
+			.forEach(v1 -> v1.setMission(null));
+
 		super.endMission(endStatus);
 	}
 
@@ -836,13 +594,11 @@ public class ConstructionMission extends AbstractMission
 	 */
 	private LightUtilityVehicle reserveLightUtilityVehicle(Settlement settlement) {
 		for(var vehicle : settlement.getParkedGaragedVehicles()) {
-			if (vehicle instanceof LightUtilityVehicle luvTemp) {
-				if (((luvTemp.getPrimaryStatus() == StatusType.PARKED) || (luvTemp.getPrimaryStatus() == StatusType.GARAGED))
+			if (vehicle instanceof LightUtilityVehicle luvTemp && ((luvTemp.getPrimaryStatus() == StatusType.PARKED) || (luvTemp.getPrimaryStatus() == StatusType.GARAGED))
 						&& !luvTemp.isReserved() && (luvTemp.getCrewNum() == 0) && (luvTemp.getRobotCrewNum() == 0)) {
-					luvTemp.setReservedForMission(true);
-					return luvTemp;
-				}
-			}
+				luvTemp.setReservedForMission(true);
+				return luvTemp;
+			}	
 		}
 
 		return null;
@@ -853,15 +609,7 @@ public class ConstructionMission extends AbstractMission
 	 * Unreserves and store back all LUV attachment parts in settlement.
 	 */
 	private void unreserveLUVparts(List<Integer> parts, Settlement settlement) {
-
-		for(Integer part : parts) {
-			try {
-				settlement.storeItemResource(part, 1);
-			} catch (Exception e) {
-				logger.severe(settlement, 5_000, 
-						"Error storing attachment part " + ItemResourceUtil.findItemResource(part).getName());
-			}
-		}
+		parts.forEach(p -> settlement.storeItemResource(p, 1));
 	}
 
 	/**
@@ -885,10 +633,8 @@ public class ConstructionMission extends AbstractMission
 		Vehicle vehicle = RoverMission.getVehicleWithGreatestRange(settlement, true);
 		if (vehicle != null) {
 			vehicleTarget = RescueSalvageVehicle.findBeaconVehicle(settlement, vehicle.getEstimatedRange());
-			if (vehicleTarget != null) {
-				if (!RescueSalvageVehicle.isClosestCapableSettlement(settlement, vehicleTarget))
-					result = true;
-			}
+			if (vehicleTarget != null && !RescueSalvageVehicle.isClosestCapableSettlement(settlement, vehicleTarget))
+				return true;
 		}
 
 		return result;
@@ -910,552 +656,6 @@ public class ConstructionMission extends AbstractMission
 	 */
 	public ConstructionStage getConstructionStage() {
 		return objective.getStage();
-	}
-
-	private static boolean positionSameBuildingType(String buildingType, double dist, ConstructionSite site) {
-		boolean goodPosition = false;
-		// Try to put building next to the same building type.
-		List<Building> sameBuildings = site.getSettlement().getBuildingManager()
-				.getBuildingsOfSameType(buildingType);
-		
-		Collections.shuffle(sameBuildings);
-		Iterator<Building> j = sameBuildings.iterator();
-		while (j.hasNext()) {
-			Building building = j.next();
-			goodPosition = positionNextToBuilding(site, building, Math.round(dist), false);
-			if (goodPosition) {
-				logger.info(site.getSettlement(), "Case 1. The building type '" 
-						+ buildingType + "' has life support.");
-				break;
-			}
-		}
-		return goodPosition;
-	}
-
-	/**
-	 * Determines and sets the position of a new construction site.
-	 *
-	 * @param site the new construction site.
-	 */
-	public static void positionNewSite(ConstructionSite site) {
-
-		boolean goodPosition = false;
-
-		Settlement s = site.getSettlement();
-		// Use settlement's objective to determine the desired building type
-		String buildingType = ObjectiveUtil.getBuildingType(s.getObjective());
-		
-		logger.info(s, "Applying building type '" + buildingType + "' as reference for '" + site + "'.");
-		
-		if (buildingType != null) {
-			
-			if (buildingConfig == null) {
-				buildingConfig = SimulationConfig.instance().getBuildingConfiguration();
-			}
-			
-			BuildingSpec spec = buildingConfig.getBuildingSpec(buildingType);
-			site.setWidth(spec.getWidth());
-			site.setLength(spec.getLength());
-			boolean isBuildingConnector = spec.getFunctionSupported().contains(FunctionType.CONNECTION);
-			boolean hasLifeSupport = spec.getFunctionSupported().contains(FunctionType.LIFE_SUPPORT);
-
-			if (isBuildingConnector) {
-				// Try to find best location to connect two buildings.
-				goodPosition = positionNewBuildingConnectorSite(site, buildingType);
-			} 
-			
-			else if (hasLifeSupport) {
-
-				goodPosition = positionSameBuildingType(buildingType, DEFAULT_FARMING_DISTANCE, site);
-		
-				if (!goodPosition) {
-					// Try to put building next to another habitable building.
-					List<Building> habitableBuildings = site.getSettlement().getBuildingManager()
-							.getBuildings(FunctionType.LIFE_SUPPORT);
-					Collections.shuffle(habitableBuildings);
-					for (Building b : habitableBuildings) {
-						// Match the floor area (e.g look more organize to put all 7m x 9m next to one
-						// another)
-						if (b.getFloorArea() == site.getWidth() * site.getLength()) {
-							goodPosition = positionNextToBuilding(site, b, DEFAULT_HABITABLE_BUILDING_DISTANCE,
-									false);
-							if (goodPosition) {
-								logger.info(s, "Case 2. Habitable.");
-								break;
-							}
-						}
-					}
-				}
-			}
-			else {
-				// Try to put building next to the same building type.
-				logger.info(s, "Case 3. Inhabitable.");
-				goodPosition = positionSameBuildingType(buildingType, DEFAULT_INHABITABLE_BUILDING_DISTANCE, site);
-			}
-		}
-
-		else {
-			// Case 4: building type is null
-			
-			// Try to put building next to another habitable building.
-			List<Building> habitableBuildings = s.getBuildingManager().getBuildings(FunctionType.LIFE_SUPPORT);
-			Collections.shuffle(habitableBuildings);
-			for (Building b : habitableBuildings) {
-				// Match the floor area (e.g look more organize to put all 7m x 9m next to one
-				// another)
-				if (b.getFloorArea() == site.getWidth() * site.getLength()) {
-					goodPosition = positionNextToBuilding(site, b, DEFAULT_INHABITABLE_BUILDING_DISTANCE, false);
-					if (goodPosition) {
-						logger.info(s, "Case 4. Habitable. Building type not given.");
-						break;
-					}
-				}
-			}
-		}
-
-		if (!goodPosition) {
-			// Try to put building next to another building.
-			// If not successful, try again 10m from each building and continue out at 10m
-			// increments
-			// until a location is found.
-			BuildingManager buildingManager = site.getSettlement().getBuildingManager();
-			if (buildingManager.getNumBuildings() > 0) {
-				for (int x = 10; !goodPosition; x += 10) {
-					for (Building b : buildingManager.getBuildingSet()) {
-						goodPosition = positionNextToBuilding(site, b, (double) x, false);
-						if (goodPosition) {
-							logger.info(s, "Case 5. Any one of the buildings.");
-							break;
-						}
-					}
-				}
-			} 
-			
-			else {
-				logger.info(s, "Case 6. No buildings found.");
-				// If no buildings at settlement, position new construction site at (0, 0) with
-				// random facing.
-				int angle = RandomUtil.getRandomInt(4) * 90;
-				site.setFacing(angle);
-				site.setPosition(LocalPosition.DEFAULT_POSITION);
-			}
-		}
-	}
-
-	/**
-	 * Determine the position and length (for variable length sites) for a new
-	 * building connector construction site.
-	 *
-	 * @param site         the construction site.
-	 * @param buildingType the new building type.
-	 * @return true if position/length of construction site could be found, false if
-	 *         not.
-	 */
-	private static boolean positionNewBuildingConnectorSite(ConstructionSite site, String buildingType) {
-
-		boolean result = false;
-
-		BuildingManager manager = site.getSettlement().getBuildingManager();
-		List<Building> inhabitableBuildings = manager.getBuildings(FunctionType.LIFE_SUPPORT);
-		Collections.shuffle(inhabitableBuildings);
-
-		if (buildingConfig == null) {
-			buildingConfig = SimulationConfig.instance().getBuildingConfiguration();
-		}
-
-		BuildingSpec spec = buildingConfig.getBuildingSpec(buildingType);
-
-		int baseLevel = spec.getBaseLevel();
-
-		// Try to find a connection between an inhabitable building without access to
-		// airlock and
-		// another inhabitable building with access to an airlock.
-		if (site.getSettlement().getAirlockNum() > 0) {
-
-			double leastDistance = Double.MAX_VALUE;
-
-			Iterator<Building> i = inhabitableBuildings.iterator();
-			while (i.hasNext()) {
-				Building startingBuilding = i.next();
-				if (!site.getSettlement().hasWalkableAvailableAirlock(startingBuilding)) {
-
-					// Find a different inhabitable building that has walkable access to an airlock.
-					Iterator<Building> k = inhabitableBuildings.iterator();
-					while (k.hasNext()) {
-						Building building = k.next();
-						if (!building.equals(startingBuilding)) {
-
-							// Check if connector base level matches either building.
-							boolean matchingBaseLevel = (baseLevel == startingBuilding.getBaseLevel())
-									|| (baseLevel == building.getBaseLevel());
-
-							if (site.getSettlement().hasWalkableAvailableAirlock(building) && matchingBaseLevel) {
-								double distance = startingBuilding.getPosition().getDistanceTo(building.getPosition());
-								if ((distance < leastDistance) && (distance >= MINIMUM_CONNECTOR_LENGTH)) {
-
-									// Check that new building can be placed between the two buildings.
-									if (positionConnectorBetweenTwoBuildings(buildingType, site, startingBuilding,
-											building)) {
-										leastDistance = distance;
-										result = true;
-									}
-								}
-							}
-						}
-					}
-				}
-			}
-		}
-
-		// Try to find valid connection location between two inhabitable buildings with
-		// no joining walking path.
-		if (!result) {
-
-			double leastDistance = Double.MAX_VALUE;
-
-			Iterator<Building> j = inhabitableBuildings.iterator();
-			while (j.hasNext()) {
-				Building startingBuilding = j.next();
-
-				// Find a different inhabitable building.
-				Iterator<Building> k = inhabitableBuildings.iterator();
-				while (k.hasNext()) {
-					Building building = k.next();
-					boolean hasWalkingPath = site.getSettlement().getBuildingConnectorManager().hasValidPath(startingBuilding,
-							building);
-
-					// Check if connector base level matches either building.
-					boolean matchingBaseLevel = (baseLevel == startingBuilding.getBaseLevel())
-							|| (baseLevel == building.getBaseLevel());
-
-					if (!building.equals(startingBuilding) && !hasWalkingPath && matchingBaseLevel) {
-
-						double distance = startingBuilding.getPosition().getDistanceTo(building.getPosition());
-						if ((distance < leastDistance) && (distance >= MINIMUM_CONNECTOR_LENGTH)) {
-
-							// Check that new building can be placed between the two buildings.
-							if (positionConnectorBetweenTwoBuildings(buildingType, site, startingBuilding, building)) {
-								leastDistance = distance;
-								result = true;
-							}
-						}
-					}
-				}
-			}
-		}
-
-		// Try to find valid connection location between two inhabitable buildings that
-		// are not directly connected.
-		if (!result) {
-
-			double leastDistance = Double.MAX_VALUE;
-
-			Iterator<Building> j = inhabitableBuildings.iterator();
-			while (j.hasNext()) {
-				Building startingBuilding = j.next();
-
-				// Find a different inhabitable building.
-				Iterator<Building> k = inhabitableBuildings.iterator();
-				while (k.hasNext()) {
-					Building building = k.next();
-					boolean directlyConnected = (site.getSettlement().getBuildingConnectorManager()
-							.getBuildingConnections(startingBuilding, building).size() > 0);
-
-					// Check if connector base level matches either building.
-					boolean matchingBaseLevel = (baseLevel == startingBuilding.getBaseLevel())
-							|| (baseLevel == building.getBaseLevel());
-
-					if (!building.equals(startingBuilding) && !directlyConnected && matchingBaseLevel) {
-						double distance = startingBuilding.getPosition().getDistanceTo(building.getPosition());
-						if ((distance < leastDistance) && (distance >= MINIMUM_CONNECTOR_LENGTH)) {
-
-							// Check that new building can be placed between the two buildings.
-							if (positionConnectorBetweenTwoBuildings(buildingType, site, startingBuilding, building)) {
-								leastDistance = distance;
-								result = true;
-							}
-						}
-					}
-				}
-			}
-		}
-
-		// Try to find connection to existing inhabitable building.
-		if (!result) {
-
-			// If variable length, set construction site length to default.
-			if (spec.getLength() == -1D) {
-				site.setLength(DEFAULT_VARIABLE_BUILDING_LENGTH);
-			}
-
-			Iterator<Building> l = inhabitableBuildings.iterator();
-			while (l.hasNext()) {
-				Building building = l.next();
-				// Make connector building face away from building.
-				result = positionNextToBuilding(site, building, 0D, true);
-
-				if (result) {
-					break;
-				}
-			}
-		}
-
-		return result;
-	}
-
-	/**
-	 * Determine the position and length (for variable length) for a connector
-	 * building between two existing buildings.
-	 *
-	 * @param buildingType   the new connector building type.
-	 * @param site           the construction site.
-	 * @param firstBuilding  the first of the two existing buildings.
-	 * @param secondBuilding the second of the two existing buildings.
-	 * @return true if position/length of construction site could be found, false if
-	 *         not.
-	 */
-	private static boolean positionConnectorBetweenTwoBuildings(String buildingType, ConstructionSite site,
-			Building firstBuilding, Building secondBuilding) {
-
-		boolean result = false;
-
-		// Determine valid placement lines for connector building.
-		List<Line2D> validLines = new ArrayList<>();
-
-		// Check each building side for the two buildings for a valid line unblocked by
-		// obstacles.
-		if (buildingConfig == null) {
-			buildingConfig = SimulationConfig.instance().getBuildingConfiguration();
-		}
-		
-		BuildingSpec spec = buildingConfig.getBuildingSpec(buildingType);
-		double width = spec.getWidth();
-		List<Point2D> firstBuildingPositions = getFourPositionsSurroundingBuilding(firstBuilding, .1D);
-		List<Point2D> secondBuildingPositions = getFourPositionsSurroundingBuilding(secondBuilding, .1D);
-		
-		for (int x = 0; x < firstBuildingPositions.size(); x++) {
-			for (int y = 0; y < secondBuildingPositions.size(); y++) {
-
-				Point2D firstBuildingPos = firstBuildingPositions.get(x);
-				Point2D secondBuildingPos = secondBuildingPositions.get(y);
-
-				double distance = Point2D.distance(firstBuildingPos.getX(), firstBuildingPos.getY(),
-						secondBuildingPos.getX(), secondBuildingPos.getY());
-
-				if (distance >= MINIMUM_CONNECTOR_LENGTH) {
-					// Check line rect between positions for obstacle collision.
-					Line2D line = new Line2D.Double(firstBuildingPos.getX(), firstBuildingPos.getY(),
-							secondBuildingPos.getX(), secondBuildingPos.getY());
-					boolean clearPath = LocalAreaUtil.isLinePathCollisionFree(line, site.getSettlement().getCoordinates(), false);
-					if (clearPath) {
-						validLines.add(new Line2D.Double(firstBuildingPos, secondBuildingPos));
-					}
-				}
-			}
-		}
-
-		if (!validLines.isEmpty()) {
-			result = isLineValid(validLines, site, firstBuilding, secondBuilding, width);
-		}
-
-		return result;
-	}
-	
-	private static boolean isLineValid(List<Line2D> validLines, ConstructionSite site,
-			Building firstBuilding, Building secondBuilding, double width) {
-		
-		// Find shortest valid line.
-		double shortestLineLength = Double.MAX_VALUE;
-		Line2D shortestLine = null;
-		Iterator<Line2D> i = validLines.iterator();
-		while (i.hasNext()) {
-			Line2D line = i.next();
-			double length = Point2D.distance(line.getX1(), line.getY1(), line.getX2(), line.getY2());
-			if (length < shortestLineLength) {
-				shortestLine = line;
-				shortestLineLength = length;
-			}
-		}
-
-		if (shortestLine == null)
-			shortestLine = validLines.get(0);
-	
-		// Create building template with position, facing, width and length for the
-		// connector building.
-		double shortestLineFacingDegrees = LocalAreaUtil.getDirection(shortestLine.getP1(), shortestLine.getP2());
-		Point2D p1 = adjustConnectorEndPoint(shortestLine.getP1(), shortestLineFacingDegrees, firstBuilding, width);
-		Point2D p2 = adjustConnectorEndPoint(shortestLine.getP2(), shortestLineFacingDegrees, secondBuilding,
-				width);
-		double centerX = (p1.getX() + p2.getX()) / 2D;
-		double centerY = (p1.getY() + p2.getY()) / 2D;
-		double newLength = p1.distance(p2);
-		double facingDegrees = LocalAreaUtil.getDirection(p1, p2);
-
-		// Provide the site the position, facing and length
-		site.setPosition(new LocalPosition(centerX, centerY));
-		site.setFacing(facingDegrees);
-		site.setLength(newLength);
-		
-		return true;
-	}
-
-
-	/**
-	 * Adjust the connector end point based on relative angle of the connection.
-	 *
-	 * @param point          the initial connector location.
-	 * @param lineFacing     the facing of the connector line (degrees).
-	 * @param building       the existing building being connected to.
-	 * @param connectorWidth the width of the new connector.
-	 * @return point adjusted location for connector end point.
-	 */
-	private static Point2D adjustConnectorEndPoint(Point2D point, double lineFacing, Building building,
-			double connectorWidth) {
-
-		double lineFacingRad = Math.toRadians(lineFacing);
-		double angleFromBuildingCenterDegrees = LocalAreaUtil
-				.getDirection(new Point2D.Double(building.getXLocation(), building.getYLocation()), point);
-		double angleFromBuildingCenterRad = Math.toRadians(angleFromBuildingCenterDegrees);
-		double offsetAngle = angleFromBuildingCenterRad - lineFacingRad;
-		double offsetDistance = Math.abs(Math.sin(offsetAngle)) * (connectorWidth / 2D);
-
-		double newXLoc = (-1D * Math.sin(angleFromBuildingCenterRad) * offsetDistance) + point.getX();
-		double newYLoc = (Math.cos(angleFromBuildingCenterRad) * offsetDistance) + point.getY();
-
-		return new Point2D.Double(newXLoc, newYLoc);
-	}
-
-	/**
-	 * Gets four positions surrounding a building with a given distance from its
-	 * edge.
-	 *
-	 * @param building         the building.
-	 * @param distanceFromSide distance (distance) for positions from the edge of
-	 *                         the building.
-	 * @return list of four positions.
-	 */
-	private static List<Point2D> getFourPositionsSurroundingBuilding(Building building, double distanceFromSide) {
-
-		List<Point2D> result = new ArrayList<>(4);
-
-		final int front = 0;
-		final int back = 1;
-		final int right = 2;
-		final int left = 3;
-
-		for (int x = 0; x < 4; x++) {
-			double xPos = 0D;
-			double yPos = 0D;
-
-			switch (x) {
-			case front:
-				xPos = 0D;
-				yPos = (building.getLength() / 2D) + distanceFromSide;
-				break;
-			case back:
-				xPos = 0D;
-				yPos = 0D - (building.getLength() / 2D) - distanceFromSide;
-				break;
-			case right:
-				xPos = 0D - (building.getWidth() / 2D) - distanceFromSide;
-				yPos = 0D;
-				break;
-			case left:
-				xPos = (building.getWidth() / 2D) + distanceFromSide;
-				yPos = 0D;
-				break;
-			}
-
-			Point2D position = LocalAreaUtil.convert2SettlementPos(xPos, yPos, building);
-			result.add(position);
-		}
-
-		return result;
-	}
-
-	/**
-	 * Positions a new construction site near an existing building.
-	 *
-	 * @param site               the new construction site.
-	 * @param building           the existing building.
-	 * @param separationDistance the separation distance (meters) from the building.
-	 * @param faceAway           true if new building should face away from other
-	 *                           building.
-	 * @return true if construction site could be positioned, false if not.
-	 */
-	private static boolean positionNextToBuilding(ConstructionSite site, Building building, double separationDistance,
-			boolean faceAway) {
-
-		boolean goodPosition = false;
-
-		final int front = 0;
-		final int back = 1;
-		final int right = 2;
-		final int left = 3;
-
-		List<Integer> directions = new ArrayList<>(4);
-		directions.add(front);
-		directions.add(back);
-		directions.add(right);
-		directions.add(left);
-		Collections.shuffle(directions);
-
-		double direction = 0D;
-		double structureDistance = 0D;
-		double rectRotation = building.getFacing();
-
-		for (int x = 0; x < directions.size(); x++) {
-			switch (directions.get(x)) {
-			case front:
-				direction = building.getFacing();
-				structureDistance = (building.getLength() / 2D) + (site.getLength() / 2D);
-				break;
-			case back:
-				direction = building.getFacing() + 180D;
-				structureDistance = (building.getLength() / 2D) + (site.getLength() / 2D);
-				if (faceAway) {
-					rectRotation = building.getFacing() + 180D;
-				}
-				break;
-			case right:
-				direction = building.getFacing() + 90D;
-				structureDistance = (building.getWidth() / 2D) + (site.getWidth() / 2D);
-				if (faceAway) {
-					structureDistance = (building.getWidth() / 2D) + (site.getLength() / 2D);
-					rectRotation = building.getFacing() + 90D;
-				}
-				break;
-			case left:
-				direction = building.getFacing() + 270D;
-				structureDistance = (building.getWidth() / 2D) + (site.getWidth() / 2D);
-				if (faceAway) {
-					structureDistance = (building.getWidth() / 2D) + (site.getLength() / 2D);
-					rectRotation = building.getFacing() + 270D;
-				}
-			}
-
-			if (rectRotation > 360D) {
-				rectRotation -= 360D;
-			}
-
-			double distance = structureDistance + separationDistance;
-			double radianDirection = Math.PI * direction / 180D;
-			LocalPosition rectCenter = building.getPosition().getPosition(distance, radianDirection);
-
-			// Check to see if proposed new site position intersects with any existing
-			// buildings
-			// or construction sites.
-			BoundedObject sitePosition = new BoundedObject(rectCenter, site.getWidth(), site.getLength(), rectRotation);
-			if (site.getSettlement().getBuildingManager().isBuildingLocationOpen(sitePosition, site)) {
-				// Set the new site here.
-				site.setPosition(rectCenter);
-				site.setFacing(rectRotation);
-				goodPosition = true;
-				break;
-			}
-		}
-
-		return goodPosition;
 	}
 	
 	@Override
