+++ mars-sim-ui/src/main/java/com/mars_sim/ui/swing/tool/mission/objectives/ConstructionPanel.java
@@ -1,6 +1,6 @@
 /*
  * Mars Simulation Project
- * ConstructionMissionCustomInfoPanel.java
+ * ConstructionPanel.java
  * @date 2023-07-24
  * @author Scott Davis
  */
@@ -16,7 +16,6 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 import javax.swing.BoundedRangeModel;
 import javax.swing.JLabel;
@@ -33,21 +32,18 @@ import com.mars_sim.core.UnitEvent;
 import com.mars_sim.core.UnitEventType;
 import com.mars_sim.core.UnitListener;
 import com.mars_sim.core.building.construction.ConstructionStage;
-import com.mars_sim.core.building.construction.ConstructionStageInfo;
-import com.mars_sim.core.building.construction.ConstructionVehicleType;
 import com.mars_sim.core.goods.Good;
 import com.mars_sim.core.goods.GoodsUtil;
 import com.mars_sim.core.mission.objectives.ConstructionObjective;
 import com.mars_sim.core.person.ai.mission.MissionEvent;
 import com.mars_sim.core.person.ai.mission.MissionListener;
-import com.mars_sim.core.resource.ItemResourceUtil;
-import com.mars_sim.core.resource.ResourceUtil;
 import com.mars_sim.core.tool.Msg;
 import com.mars_sim.ui.swing.MainDesktopPane;
 import com.mars_sim.ui.swing.StyleManager;
 import com.mars_sim.ui.swing.components.EntityLabel;
 import com.mars_sim.ui.swing.tool.mission.ObjectivesPanel;
 import com.mars_sim.ui.swing.utils.AttributePanel;
+import com.mars_sim.ui.swing.utils.ConstructionStageFormat;
 
 /**
  * A panel for displaying construction custom mission information.
@@ -99,7 +95,7 @@ public class ConstructionPanel extends JPanel implements MissionListener, Object
         processPanel.add(progressBarPanel);
 
         // Add tooltip.
-        processPanel.setToolTipText(getToolTipString(stage));
+        processPanel.setToolTipText(ConstructionStageFormat.getTooltip(stage));
         
         JLabel progressLabel = new JLabel("Site Completion", SwingConstants.CENTER);
         processPanel.add(progressLabel);
@@ -180,63 +176,10 @@ public class ConstructionPanel extends JPanel implements MissionListener, Object
         progressBarModel.setValue(workProgress);
         
         // Update the tool tip string.
-        processPanel.setToolTipText(getToolTipString(stage));
+        processPanel.setToolTipText(ConstructionStageFormat.getTooltip(stage));
     }
 
-    /**
-     * Gets a tool tip string for the panel.
-     */
-    private String getToolTipString(ConstructionStage stage) {
-        StringBuilder result = new StringBuilder(Msg.HTML_START);
-
-        if (stage != null) {
-            ConstructionStageInfo info = stage.getInfo();
-            result.append("Status: building ").append(info.getName()).append(Msg.BR);
-            result.append("Stage Type: ").append(info.getType()).append(Msg.BR);
-            result.append("Work Type: Construction").append(Msg.BR);
-            String requiredWorkTime = StyleManager.DECIMAL_PLACES1.format(stage.getRequiredWorkTime() / 1000D);
-            result.append("Work Time Required: ").append(requiredWorkTime).append(" Sols").append(Msg.BR);
-            String completedWorkTime = StyleManager.DECIMAL_PLACES1.format(stage.getCompletedWorkTime() / 1000D);
-            result.append("Work Time Completed: ").append(completedWorkTime).append(" Sols").append(Msg.BR);
-            result.append("Architect Construction Skill Required: ").append(info.getArchitectConstructionSkill()).append(Msg.BR);
-
-            // Add remaining construction resources.
-            if (!stage.getMissingResources().isEmpty()) {
-                result.append(Msg.BR).append("Missing Construction Resources:").append(Msg.BR);
-                result.append(stage.getMissingResources().entrySet().stream()
-                        .map(e -> Msg.NBSP + ResourceUtil.findAmountResourceName(e.getKey()) + ": " + e.getValue()
-                                        + " kg")
-                        .collect(Collectors.joining(Msg.BR)));
-            }
-
-            // Add Missing construction parts.
-            if (!stage.getMissingParts().isEmpty()) {
-                result.append(Msg.BR).append("Missing Construction Parts:").append(Msg.BR);
-                result.append(stage.getMissingParts().entrySet().stream()
-                    .map(e -> Msg.NBSP + ItemResourceUtil.findItemResourceName(e.getKey()) + ": " + e.getValue())
-                    .collect(Collectors.joining(Msg.BR)));
-            }
-
-            // Add construction vehicles.
-            if (!info.getVehicles().isEmpty()) {
-                result.append(Msg.BR).append("Construction Vehicles").append(Msg.BR);
-                
-                for(ConstructionVehicleType vehicle : info.getVehicles()) {
-                    result.append(Msg.NBSP).append(Msg.NBSP).append("Vehicle Type: ").append(vehicle.getVehicleType()).append(Msg.BR);
-                    result.append(Msg.NBSP).append(Msg.NBSP).append("Attachment Parts:").append(Msg.BR);
-                    Iterator<Integer> l = vehicle.getAttachmentParts().iterator();
-                    while (l.hasNext()) {
-                        result.append(Msg.NBSP).append(Msg.NBSP).append(Msg.NBSP).append(Msg.NBSP)
-                        .append("-").append(ItemResourceUtil.findItemResourceName(l.next())).append(Msg.BR);
-                    }
-                }
-            }
-        }
-
-        result.append(Msg.HTML_STOP);
-
-        return result.toString();
-    }
+  
 
     /**
      * Model for the construction materials table.
